pyproject.toml
[project]
name = "coord2loc"
version = "0.0.3q"
description = "Convert latitude and longitude into localities like countries and states"
readme = "README.md"
authors = [
    { name = "James Hale", email = "james+coord2loc@hale.zip" }
]
requires-python = ">=3.9"
dependencies = [
    "msgpack>=1.1.0",
    "shapely>=2.0.7",
]

[project.entry-points."coord2loc.data"]

[project.optional-dependencies]
all = ["coord2loc-can", "coord2loc-usa"]
usa = ["coord2loc-usa"]
can = ["coord2loc-can"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
build = [
    "geopandas>=1.0.1",
]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/coord2loc/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/coord2loc"]

[tool.ruff.lint]
extend-select = ["I", "RUF022"]

[tool.uv.sources]
coord2loc = { workspace = true }
coord2loc-can = { workspace = true }
coord2loc-usa = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]
